Create user management console
Add User managerment to admin section. Add Angular-bootstrap library for modals.
Create UserService Angular factory to interact with Sling UserManager Post
servlets.

There are three groups for users. "authors" for blog content publishers. "testers"
to log in and view the site, but not make changes - mainly for demo purposes. The
third "group" isn't really a group. It's called "admin" and includes users without
a group such as the default "admin" and "anonymous". Users can be added "authors"
and "testers" but not admin, and same for deletions. Any user can be updated in
that their display name or password can be changed.