Update Apache redirects for Sling 7 bug
The post servlet that handles updating user groups doesn't work when
posting to JSON in Sling 7. It has been fixed in Sling 8, however,
this project is using the latest version of Sling's downloadable JAR,
which is 7. The workaround is for the UserService in admin.js to post
to HTML rather than JSON to update groups. The side effect is that in
order to use extensionless URLs, you need to account for this in your
web server redirects.