OSGi config service & implement on System Settings
Create OSGi service to set and read configurations. Implement the service in the
SystemSettingsService rather than saving to the /etc/config node. Update the
System Config Servlet to check for authorable users and return a 403 error if not
authorized to make changes. Update JavaScript to ensure proper error code is
displayed.