Implement OSGi config service for reCAPTCHA settings
Update recaptcha servlet to save settings through the service rather than writing
to a node itself. Update the recaptcha service to add setter methods and use the
OSGi service to write to the OSGi configurations instead of the /etc/config node.

Update comments on SystemSettingsService to match comments on the Recaptcha
Service.