Add Akismet spam filtering service
Create Akismet config component. Add component to a new admin config page and add
the page to the config section of the dashboard. Wire the component up with the
Angular settings controller, update the JS settings service to post to a new
servlet. Create the Akismet service to get and set Akismet configuration
settings. Create Akismet service methods to check comments, submit spam, and
submit ham. Create an HttpService with reusable helper methods for getting the
User Agent, IP Address, and referrer from requests. Include the java-akismet
bundle in core as it isn't an OSGi bundle. Include dependencies for the java-
akismet artifact.