Add admin comments section
Create comments page in the admin section and add a commentList control. Create
a commentList control that uses Angular to get all comments. Create an Angular
comment controller and comment service to communicate with the server. Add the
new JS files to the Grunt build task. Create a servlet to respond to the GET
request for all comments. Create a comment service to provide comment CRUD.

TODO: move comment creation from current servlet to the service. Create client
side work to tie in deletion, editing, spam and ham updates using the servlet
and Akismet service.