Add admin ability to mark comments as spam or ham
Add Spam and Ham functions in the comment servlet, which calls the Comment
Service. Add markSpam() and markHam() methods to the Comment Service which calls
the Akamai Service. Update the Akamai services submitSpam and submitHam methods
to accept the comment node ID and update spam and display properties.

Add client-side Akismet functionality to the comment list admin component. Update
the Comment Controller, Comment Modal Controller, and Comment Service to open a
modal, communicate with the server and update the Angular model upon success.
Rearrange the comment list view table.

By marking a comment as spam it be hidden and submitted as spam to Akismet. By
marking a comment as ham it will be shown and submitted as ham to Akismet. If the
user submits spam, it will be hidden to the user but visible to the author with
the option of marking it as ham.

TODO: refactor comment view to use service.
TODO: refactor comment servlet to use service and AJAX.