Filter blog posts that aren't published
If the author hasn't checked the "show blog post" checkbox, then the post should
be in preview mode and not display to the general public. Anonymous users get a
404 error if the post isn't published. Admins, authors and testers can view the
unpublished page but a message is injected into the top of the page stating that
the post isn't live.

Include fake requests and responses from the Apache Sling source. Create the
message as a component. Add the functionality to the preview filter.

TODO: Can the SlingRequestProcessor return the HTML of the component without
wrapping it in extra HTML?

TODO: Replace the filter and HTML injection by simply Sling Including the
component through the blogPage.