Add Gist embedding
Add gist-embed jQuery plugin and include it on the public pages. Create plugin
for Summernote WYSIWYG editor and add it to the toolbar. Add XSS anti-sammy
config file and allow data-gist-* attributes. Update CSS for admin to show the
gist ID in the editor. Update CSS for public to not surround the gist code
elements in a well. Update Sightly and Sling XSS from 1.0.2 to 1.0.4.

Note: It seems you have to restart the Sling server for the XSS config to take
effect.