Display links to all blog posts from every post
Create the Collapsible Blog List component which is a Bootstrap based Sightly
component in the form of an accordion-like widget. The user can drill down to
each blog post by clicking on the year, then the month, then the blog post.

Add the new component to the blog view component in the right rail.

Move the comment component from the page to the blog view in order to facilitate
responsive wrapping where the comments appear above the right rail.